+++ CHANGELOG.md
@@ -11,9 +11,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Dropped "loading-indicator--dot" component style from LoadingIndicator https://github.com/Textualize/textual/pull/2050
 
+### Removed
+
+- Removed `sender` attribute from messages. It's now just private (`_sender`). https://github.com/Textualize/textual/pull/2071
+
 ### Fixed
 
 - Fixed borders not rendering correctly. https://github.com/Textualize/textual/pull/2074
+
 
 ### Changed